+++ packages/plugin-divider/web/src/icons/DoubleLineStyle.tsx
@@ -5,7 +5,7 @@ const DoubleLineStyle = props => (
     width="118"
     height="12"
     viewBox="0 0 118 6"
-    fill="none"
+    fill="currentColor"
     xmlns="http://www.w3.org/2000/svg"
     {...props}
   >