+++ packages/plugin-divider/web/src/icons/SingleLineStyle.tsx
@@ -5,7 +5,7 @@ const SingleLineStyle = props => (
     width="118"
     height="1"
     viewBox="0 0 118 1"
-    fill="none"
+    fill="currentColor"
     xmlns="http://www.w3.org/2000/svg"
     {...props}
   >