+++ packages/plugin-table/web/src/toolbar/buttons/VerticalAlignmentButton.tsx
@@ -26,6 +26,7 @@ const VerticalAlignmentButton: FC<Props> = ({ dataHook, toolbarItem }) => {
   const modalService = useContext(ModalContext) || {};
 
   const [referenceElement, setReferenceElement] = useState<HTMLDivElement | null>(null);
+  const [isButtonActive, setIsButtonActive] = useState(false);
 
   const verticalAlignment = toolbarItem.attributes.cellAlignment || 'TOP';
   const Icon = iconsMap[verticalAlignment];
@@ -52,9 +53,18 @@ const VerticalAlignmentButton: FC<Props> = ({ dataHook, toolbarItem }) => {
       : closeModal();
   };
 
+  modalService.onModalOpened(() => {
+    modalService.isModalOpen(TABLE_BUTTONS_MODALS_ID.VERTICAL_ALIGNMENT) && setIsButtonActive(true);
+  });
+  modalService.onModalClosed(() => {
+    !modalService.isModalOpen(TABLE_BUTTONS_MODALS_ID.VERTICAL_ALIGNMENT) &&
+      setIsButtonActive(false);
+  });
+
   return (
     <DropdownButton
       dataHook={dataHook}
+      active={isButtonActive}
       onClick={openVerticalAlignmentPanel}
       setRef={setReferenceElement}
       Icon={Icon}