+++ packages/toolbars-ui/web/src/buttons/NodeSizeButton/NodeSizeButton.tsx
@@ -15,6 +15,7 @@ const NodeSizeButton: FC<Props> = ({ toolbarItem, dataHook, id }) => {
   const { t, isMobile } = useContext(RicosContext) || {};
   const modalService = useContext(ModalContext) || {};
   const [referenceElement, setReferenceElement] = useState<HTMLDivElement | null>(null);
+  const [isButtonActive, setIsButtonActive] = useState(false);
   const getSelectedSize: () => string = () => toolbarItem?.attributes.nodeSize || 'CONTENT';
   const selectedSize: string = getSelectedSize();
   const SelectedSizeIcon = sizeIconMap[`${selectedSize}`];
@@ -38,9 +39,17 @@ const NodeSizeButton: FC<Props> = ({ toolbarItem, dataHook, id }) => {
         });
   };
 
+  modalService.onModalOpened(() => {
+    modalService.isModalOpen(id) && setIsButtonActive(true);
+  });
+  modalService.onModalClosed(() => {
+    !modalService.isModalOpen(id) && setIsButtonActive(false);
+  });
+
   return (
     <DropdownButton
       dataHook={dataHook}
+      active={isButtonActive}
       onClick={onClick}
       setRef={setReferenceElement}
       Icon={SelectedSizeIcon}